Allow the user to configure the position update interval
After I started using mpd-mpris on my laptop I noticed significantly
more wakeups than before even when mpd is not playing anything. This
seemed surprising, because applications using dbus typically do not
require polling.

Looking more closely, the polling comes from updating the current song
position, since this doesn't generate dbus events. As such, this also
happens even when mpd is idle.

The interval is now configurable, and if set to 0, never updates the
status just for position information (although, of course, it will be
updated on other changes that result in a status update).

On my machine, this renders mpd-mpris significantly more willing to
sleep and avoid wakeups, especially since I don't use this feature
anyway, so it's nice to be able to disable it with `--interval 0`.