Use `Name` attribute as fallback (#36)
In some situations the `Title` attribute is missing (e.g. poorly
configured internet radio stations).  In that case, we can fallback
on the `Name` attribute and, if that fails, the `file` attribute.

`xesam:name` does not seem to be common[1], so using `Title`
instead of adding a `Name` to the `File` struct.

[1] https://www.freedesktop.org/wiki/Specifications/mpris-spec/metadata/

Co-authored-by: André Marçais <pex@marcais.net>