Fix song refresh when playing audio stream
Audio streams (e.g. web radios) are always the same file (i.e. the m3u
file). However the song will change over time as sent by the stream.

We can't really know if the file is a stream or not, so we suppose it's
a stream when the duration is 0.

Fixes #57

Signed-off-by: Frank Villaro-Dixon <frank@villaro-dixon.eu>